Accept Lighter testnet market ids above 255.
Testnet BTC/ETH now use ids like 4096, which the bot was dropping as invalid.

+++ backend/cmd/bot/main.go
@@ -90,7 +90,7 @@ func main() {
 		os.Exit(1)
 	}
 
-	idToSym := map[uint8]string{}
+	idToSym := map[uint16]string{}
 	for s, m := range markets {
 		idToSym[m.MarketID] = s
 	}
